requirements.txt
# requirements.txt - Python 3.13 Compatible Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0

# Redis (with fallback)
redis>=5.0.1
aioredis>=2.0.1

# HTTP and async
aiohttp>=3.9.1
aiofiles>=23.2.0
httpx>=0.25.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0